refactor(create): replace site nav with builder top bar on /create
